requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Django==4.2.15
django-oauth-toolkit==2.4.0
django-rest-knox==5.0.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
idna==3.7
jwcrypto==1.5.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
